refactor: checkForWin logic
This commit adds a checkForDraw method to account for draw games.
I have also refactored the code to fix the previous eslint conflicts,
where the code was using a variable in the playRound method before it was declared.
Also refactored the for of loop in checkForWin which was also causing an eslint warning.