add: board reset functionality
This update creates a reset board function and connects it to the html reset__button element.
Allows players to restart the game after a win, draw or even mid way through a game.

ESlint currently warning against using a unary operator within the nested for loop.
I will look to resolve this issue in a future update.