fix: eslint warnings
The use of unary operators in the resetBoard function were being flagged by eslint.
However I don't think this is of concern given the scope of this project so have disabled it.

Additionally I renamed the row and col variables to clickRow and clickCol as these names were already being used in line 96:
const [row, col] = getRowCol(index);

This update fixes the outstanding eslint warnings.